==> code/02_data_cleaning_and_eda.ipynb <==
# Project 3: Web APIs & Classification

## Notebook 2: Data Cleaning and EDA

### Contents:
- [EDA](#EDA)
- [Create classification column and combine data](#Create-classification-column-and-combine-data)

import pandas as pd
import ast

dems_uncleaned_df = pd.read_csv('../datasets/dems_top1000_posts.csv')
reps_uncleaned_df = pd.read_csv('../datasets/reps_top1000_posts.csv')

### EDA

# check to ensure the files are in the correct shape
print(dems_uncleaned_df.shape, reps_uncleaned_df.shape)

dems_uncleaned_df['data'][0]

type(dems_uncleaned_df['data'][0])

We can see that the data which is originally a dictionary has now been changed to a string type hence we would need to change it back into a string so that we can extract the relevant key value pair. We do it using the `ast` library function `literal_eval`. 

dems_uncleaned_df['data'] = dems_uncleaned_df['data'].map(lambda data: ast.literal_eval(data))
reps_uncleaned_df['data'] = reps_uncleaned_df['data'].map(lambda data: ast.literal_eval(data))

# check that it has successfully changed to a dict type
type(dems_uncleaned_df['data'][0])

# the reddit post title
dems_uncleaned_df['data'][0]['title']

# the reddit post text
dems_uncleaned_df['data'][0]['selftext']

# the reddit post id
dems_uncleaned_df['data'][0]['name']

# the reddit post upvotes
dems_uncleaned_df['data'][0]['ups']

# the reddit post downvotes
dems_uncleaned_df['data'][0]['downs']

len(dems_uncleaned_df)

# iterate through the dataset and extract these 5 key value pair
dems_titles = [dems_uncleaned_df['data'][i]['title'] for i in range (len(dems_uncleaned_df))]
dems_selftext = [dems_uncleaned_df['data'][i]['selftext'] for i in range (len(dems_uncleaned_df))]
dems_name = [dems_uncleaned_df['data'][i]['name'] for i in range (len(dems_uncleaned_df))]
dems_ups = [dems_uncleaned_df['data'][i]['ups'] for i in range (len(dems_uncleaned_df))]
dems_downs = [dems_uncleaned_df['data'][i]['downs'] for i in range (len(dems_uncleaned_df))]

print(len(dems_titles), len(dems_selftext), len(dems_name), len(dems_ups), len(dems_downs))

# convert the data into a pandas dataframe
dems_df = pd.DataFrame([dems_name, dems_titles, dems_selftext, dems_ups, dems_downs],
                       index = ['post_id', 'title', 'post_text', 'upvotes', 'downvotes']).T

# the same process is done on the republican dataset
reps_titles = [reps_uncleaned_df['data'][i]['title'] for i in range (len(reps_uncleaned_df))]
reps_selftext = [reps_uncleaned_df['data'][i]['selftext'] for i in range (len(reps_uncleaned_df))]
reps_name = [reps_uncleaned_df['data'][i]['name'] for i in range (len(reps_uncleaned_df))]
reps_ups = [reps_uncleaned_df['data'][i]['ups'] for i in range (len(reps_uncleaned_df))]
reps_downs = [reps_uncleaned_df['data'][i]['downs'] for i in range (len(reps_uncleaned_df))]

print(len(reps_titles), len(reps_selftext), len(reps_name), len(reps_ups), len(reps_downs))

reps_df = pd.DataFrame([reps_name, reps_titles, reps_selftext, reps_ups, reps_downs],
                       index = ['post_id', 'title', 'post_text', 'upvotes', 'downvotes']).T

### Create classification column and combine data

We would need to combine the dataframe but before we do that we must designate a binary value to seperate dataframe, this new column will be `'is_dems'` where `1` is for all posts from r/democrats and `0` is for all posts from r/Republican

dems_df['is_dems'] = 0
reps_df['is_dems'] = 1

# we use the ignore_index to ensure that the index is reset as we do not need to preserve it
df = pd.concat([dems_df, reps_df], ignore_index = True)

df.head()

df.tail()

df.shape

# check if there is any null value
df.isnull().sum()

df['is_dems'].value_counts()

post_with_text = [post for post in df['post_text'] if (len(post) != 0)]

len(post_with_text)

df['combine_text'] = df['title'] + ' ' + df['post_text']

df.head()

df.to_csv('../datasets/df_eda.csv', index = False)

## Continue to Notebook 03
