==> code/04_modeling_evaluation_conclusion_and_recommendation.ipynb <==
# Project 3: Web APIs & Classification

## Notebook 4: Modeling Evaluation Conclusion and Recommendation

### Contents:
- [Model selections:](#Model-selections:)
- [Multinomial Naive-Bayes](#Multinomial-Naive-Bayes)
- [Logistic Regressions](#Logistic-Regressions)
- [Random Forest](#Random-Forest)
- [Evaluating Models](#Evaluating-Models)
- [Conclusion and Recommendations](#Conclusion-and-Recommendations)

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import itertools

from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.model_selection import GridSearchCV, train_test_split, cross_val_score
from sklearn.pipeline import Pipeline
from sklearn.preprocessing import StandardScaler
from sklearn.linear_model import LogisticRegression
from sklearn.naive_bayes import MultinomialNB
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import confusion_matrix, accuracy_score, precision_score, recall_score, f1_score 

X_train = pd.read_csv('../datasets/X_train.csv', index_col = 0)
X_test = pd.read_csv('../datasets/X_test.csv', index_col = 0)
y_train = pd.read_csv('../datasets/y_train.csv', index_col = 0)
y_test = pd.read_csv('../datasets/y_test.csv', index_col = 0)

X_train.head()

X_train_final = X_train['combine_text_tokens'].values.astype('U')
X_test_final = X_test['combine_text_tokens'].values.astype('U')

### Model selections: 
#### 1.  TF-IDF Vectorizer Multinomial Naive-Bayes
  - `tf__ngram_range = (1, 2)`
  - `tf__stop_words = 'english'`
  
#### 2. TF-IDF Vectorizer Scaled Logistic Regression
  - `tf__ngram_range = (1, 2)`
  - `tf__stop_words = 'english'`
  
#### 3. TF-IDF Vectorizer Random Forest 
  - `tf__ngram_range = (1, 1)`
  - `tf__stop_words = 'english'`

### Multinomial Naive-Bayes

mnb_loop = pd.DataFrame(columns = ['train_accuracy', 'test_accuracy', 'best_params', 'tn', 'fp', 'fn', 'tp',
                                   'precision', 'recall', 'f1'])

mnb_steps = [('tf', TfidfVectorizer(stop_words = 'english', ngram_range = (1, 2))),
             ('mnb', MultinomialNB())]

mnb_params = {"mnb__alpha": np.arange(.1, 1, .1),
              "tf__max_features": [30000, 32500, 35000, 37500]}

mnb_params_1 = {"mnb__alpha": np.arange(.01, 1, .1),
                "tf__max_features": [22500, 25000, 27500, 30000, 32500]}

mnb_params_2 = {"mnb__alpha": np.arange(.01, 0.5, .05),
                "tf__max_features": [22500, 25000, 27500, 30000, 32500]}

pipe = Pipeline(mnb_steps)

def mnb_pipe(params):
    mnb_results = {}

    grid = GridSearchCV(pipe, params, cv = 5) # optimize GridSearch hyperparameters on `cv=5` cross validation runs
    grid.fit(X_train_final, y_train.values.ravel()) # fit to our training data

    print('Train Accuracy: ', grid.score(X_train_final, y_train))
    mnb_results['train_accuracy'] = grid.score(X_train_final, y_train) # print/store training accuracy

    print('Test Accuracy: ',grid.score(X_test_final, y_test))
    mnb_results['test_accuracy'] = grid.score(X_test_final, y_test) # print/store test accuracy

    print('Best Params: ',grid.best_params_, '\n')
    mnb_results['best_params'] = grid.best_params_ # print/store best parameters

    tn, fp, fn, tp = confusion_matrix(y_test, grid.predict(X_test_final)).ravel() # inspect counted results in matrix
    print("True Negatives: %s" % tn)
    mnb_results['tn'] = tn
    print("False Positives: %s" % fp)
    mnb_results['fp'] = fp
    print("False Negatives: %s" % fn)
    mnb_results['fn'] = fn
    print("True Positives: %s" % tp, '\n')
    mnb_results['tp'] = tp

    print("Precision Score: ", precision_score(y_test, grid.predict(X_test_final)))
    mnb_results['precision'] = precision_score(y_test, grid.predict(X_test_final))

    print("Recall Score: ", recall_score(y_test, grid.predict(X_test_final)))
    mnb_results['recall'] = recall_score(y_test, grid.predict(X_test_final))

    print("F1 Score: ", f1_score(y_test, grid.predict(X_test_final)), '\n')
    mnb_results['f1'] = f1_score(y_test, grid.predict(X_test_final))
    
    return mnb_results

mnb_results = mnb_pipe(mnb_params)

mnb_loop = mnb_loop.append(mnb_results, ignore_index = True)

mnb_results = mnb_pipe(mnb_params_1)

mnb_loop = mnb_loop.append(mnb_results, ignore_index = True)

mnb_results = mnb_pipe(mnb_params_2)

mnb_loop = mnb_loop.append(mnb_results, ignore_index = True)

mnb_loop

### Logistic Regressions

lr_loop = pd.DataFrame(columns = ['train_accuracy', 'test_accuracy', 'best_params', 'tn', 'fp', 'fn', 'tp',
                                  'precision', 'recall', 'f1'])

lr_steps = [('tf', TfidfVectorizer(stop_words = 'english', ngram_range = (1, 2))),
            ('sc', StandardScaler(with_mean = False)),
            ('lr', LogisticRegression(random_state = 42))]

lr_params = {"lr__penalty": ['l2'], "lr__C": [1.2],
             "lr__tol": [.00035], "tf__max_features": [25000, 27500, 30000, 32500, 35000]}

lr_params_1 = {"lr__penalty": ['l2'], "lr__C": [1.2],
               "lr__tol": [.00035], "tf__max_features": [15000, 17500, 20000, 22500, 25000, 27500]}

lr_params_2 = {"lr__penalty": ['l1'], "lr__C": [1, 1.2], "lr__solver": ['liblinear'],
               "lr__tol": [.00035], "tf__max_features": [14000, 15000, 16000]}

lr_params_3 = {"lr__penalty": ['l1'], "lr__C": [1, 1.2], "lr__solver": ['liblinear'],
               "lr__tol": [.00035], "tf__max_features": [12000, 13000, 14000, 15000, 16000]}

pipe = Pipeline(lr_steps)

def lr_pipe(params):
    lr_results = {}

    grid = GridSearchCV(pipe, params, cv = 5) # optimize GridSearch hyperparameters on `cv=5` cross validation runs
    grid.fit(X_train_final, y_train.values.ravel()) # fit to our training data

    print('Train Accuracy: ', grid.score(X_train_final, y_train))
    lr_results['train_accuracy'] = grid.score(X_train_final, y_train) # print/store training accuracy

    print('Test Accuracy: ',grid.score(X_test_final, y_test))
    lr_results['test_accuracy'] = grid.score(X_test_final, y_test) # print/store test accuracy

    print('Best Params: ',grid.best_params_, '\n')
    lr_results['best_params'] = grid.best_params_ # print/store best parameters

    tn, fp, fn, tp = confusion_matrix(y_test, grid.predict(X_test_final)).ravel() # inspect counted results in matrix
    print("True Negatives: %s" % tn)
    lr_results['tn'] = tn
    print("False Positives: %s" % fp)
    lr_results['fp'] = fp
    print("False Negatives: %s" % fn)
    lr_results['fn'] = fn
    print("True Positives: %s" % tp, '\n')
    lr_results['tp'] = tp

    print("Precision Score: ", precision_score(y_test, grid.predict(X_test_final)))
    lr_results['precision'] = precision_score(y_test, grid.predict(X_test_final))

    print("Recall Score: ", recall_score(y_test, grid.predict(X_test_final)))
    lr_results['recall'] = recall_score(y_test, grid.predict(X_test_final))

    print("F1 Score: ", f1_score(y_test, grid.predict(X_test_final)), '\n')
    lr_results['f1'] = f1_score(y_test, grid.predict(X_test_final))
    
    return lr_results

lr_results = lr_pipe(lr_params)

lr_loop = lr_loop.append(lr_results, ignore_index = True)

lr_results = lr_pipe(lr_params_1)

lr_loop = lr_loop.append(lr_results, ignore_index = True)

lr_results = lr_pipe(lr_params_2)

lr_loop = lr_loop.append(lr_results, ignore_index = True)

lr_loop

### Random Forest

rf_loop = pd.DataFrame(columns = ['train_accuracy', 'test_accuracy', 'best_params', 'tn', 'fp', 'fn', 'tp',
                                  'precision', 'recall', 'f1'])

rf_steps = [('tf', TfidfVectorizer(stop_words = 'english', ngram_range = (1, 1))),
            ('rf', RandomForestClassifier(random_state = 42))]

rf_params = {"rf__n_estimators": np.arange(98, 102, 1), "rf__max_depth": [7, 8, 9], 
             "rf__criterion": ['gini'], "tf__max_features": [None, 35000, 40000, 45000]}

rf_params_1 = {"rf__n_estimators": np.arange(94, 98, 1), "rf__max_depth": [5, 6, 7], 
               "rf__criterion": ['gini'], "tf__max_features": [None, 2500, 5000, 7500]}

rf_params_2 = {"rf__n_estimators": np.arange(100, 104, 1), "rf__max_depth": [6, 7, 8, 9], 
               "rf__criterion": ['gini'], "tf__max_features": [None]}

pipe = Pipeline(rf_steps)

def rf_pipe(params):
    rf_results = {}

    grid = GridSearchCV(pipe, params, cv = 5) # optimize GridSearch hyperparameters on `cv=5` cross validation runs
    grid.fit(X_train_final, y_train.values.ravel()) # fit to our training data

    print('Train Accuracy: ', grid.score(X_train_final, y_train))
    rf_results['train_accuracy'] = grid.score(X_train_final, y_train) # print/store training accuracy

    print('Test Accuracy: ',grid.score(X_test_final, y_test))
    rf_results['test_accuracy'] = grid.score(X_test_final, y_test) # print/store test accuracy

    print('Best Params: ',grid.best_params_, '\n')
    rf_results['best_params'] = grid.best_params_ # print/store best parameters

    tn, fp, fn, tp = confusion_matrix(y_test, grid.predict(X_test_final)).ravel() # inspect counted results in matrix
    print("True Negatives: %s" % tn)
    rf_results['tn'] = tn
    print("False Positives: %s" % fp)
    rf_results['fp'] = fp
    print("False Negatives: %s" % fn)
    rf_results['fn'] = fn
    print("True Positives: %s" % tp, '\n')
    rf_results['tp'] = tp

    print("Precision Score: ", precision_score(y_test, grid.predict(X_test_final)))
    rf_results['precision'] = precision_score(y_test, grid.predict(X_test_final))

    print("Recall Score: ", recall_score(y_test, grid.predict(X_test_final)))
    rf_results['recall'] = recall_score(y_test, grid.predict(X_test_final))

    print("F1 Score: ", f1_score(y_test, grid.predict(X_test_final)), '\n')
    rf_results['f1'] = f1_score(y_test, grid.predict(X_test_final))
    
    return rf_results

rf_results = rf_pipe(rf_params)

rf_loop = rf_loop.append(rf_results, ignore_index = True)

rf_loop

rf_results = rf_pipe(rf_params_1)

rf_loop = rf_loop.append(rf_results, ignore_index = True)

rf_results = rf_pipe(rf_params_2)

rf_loop = rf_loop.append(rf_results, ignore_index = True)

rf_loop

### Evaluating Models

**Accuracy** - How many posts (r/democrats or r/Republican) did we correctly predict out of all the posts? (TP + TN / TP + FP + TN + FN)    
**Precision** - How many posts of those we labeled as democrats are actually democrats post? (TP / TP + FP)  
**Recall/Sensitivity** - Of all the democrats posts how many did we correctly predict? (TP / TP + FN)     
**F1-score** - It is the harmonic mean of precision and recall (2 * (Recall * Precision) / (Recall + Precision))  

mnb_loop.sort_values('test_accuracy', ascending = False)

Index 1 is choosen out of all the Multinomial Naive-Bayes models apart from looking at the test accuracy score, it is important to also evaluate both the precision and recall of the model we can see that it also out performance the other runs by score the highest F1-score.

lr_loop.sort_values('test_accuracy', ascending = False)

Index 0 is choosen out of all the Logistic Regression models although it does not have the highest test accuracy score, upon further inspection of the precision, recall and F1-score we can see that it actually outperforms the higher test accuracy score of Index 2 model. We decided to pick the model with a higher F1-score as this means it is capable of getting a more accuracy prediction for post on democrats subreddit which is our focus.

rf_loop.sort_values('test_accuracy', ascending = False)

Index 1 scored higher compared to index 2 in terms of the test accuracy score, however, between the two index 2 has a higher F1-socre as it means it is capable of getting a more accurate prediction for post on democrats subreddict which is our focus. Hence index 2 final parameters is choosen.

mnb_loop[2:3]

final_model = pd.concat([mnb_loop[1:2], lr_loop[:1], rf_loop[2:3]], axis = 0)

final_model.sort_values('test_accuracy', ascending = False)

Looking at the final models selected for the individual model types, Random Forest model is first dropped. We then look to evaluate the Multinomial Naive-Bayes model and Logistic Regressions model. Looking all the metrics for comparison we can see that the Multinomial Naive-Bayes model outperforms the Logistic Regressions model. Hence, our final selection will be the Multinomial Naive-Bayes model with `alpha: 0.21` and `max_features = 22500`.

final_steps = [('tf', TfidfVectorizer(stop_words = 'english', ngram_range = (1, 2), max_features = 22500)),
               ('mnb', MultinomialNB(alpha = 0.21))]

pipe = Pipeline(final_steps)

pipe.fit(X_train_final, y_train.values.ravel())

pipe.score(X_train_final, y_train)

pipe.score(X_test_final, y_test)

selected_model = pipe.named_steps['mnb']

selected_vec = pipe.named_steps['tf']

selected_vec.fit_transform(X_train_final)

selected_model_df = pd.DataFrame(selected_model.coef_.T, index = selected_vec.get_feature_names(), columns = ['coef'])

selected_model_df.coef.sort_values(ascending=False).head(10)

# Function researched and borrowed from Stackoverflow
# https://stackoverflow.com/questions/11116697/how-to-get-most-informative-features-for-scikit-learn-classifiers

def important_features(vectorizer, classifier, n = 20):
    class_labels = classifier.classes_
    feature_names = vectorizer.get_feature_names()
    topn_class1 = sorted(zip(classifier.feature_count_[0], feature_names), reverse = True)[:n]
    topn_class2 = sorted(zip(classifier.feature_count_[1], feature_names), reverse = True)[:n]
    print("Important words for r/Republican\n")
    for coef, feat in topn_class1:
        print(class_labels[0], coef, feat)
    print("-----------------------------------------\n")
    print("Important words for r/democrats\n")
    for coef, feat in topn_class2:
        print(class_labels[1], coef, feat) 

important_features(selected_vec, selected_model, 15)

### Conclusion and Recommendations

Based on our findings, we were able to answer the questions that we have formulated as part of the problem statement.  
- Between the two similar subreddits r/democrats and r/Republican are able to differentiate the post using Natural Language Processing models?    

Yes we were able to differentiate the post between the two different subreddits and compared to our baseline score of 50.025% we were able to improve it to 74.6%. However, due to the overlap nature of some major key words this has likely reduce the performance of our models.  
- Which models is then likely to work best?  

Multinomial Naive-Bayes model seems to works the best after evaluating the metrics for accuracy, precision, recall and f1-score across the 3 different models that were utilised. 

While we have managed to answer our problem statement that we have formulated, there is still room for improvement. Particularly in terms of the final selected model there is a huge different between the train and test score this is likely due to overfitting. We can take a few steps to further improve the model

1. Increase data size of collection by utilising pushshift.io which is an alternative reddit API that is capable of exceeding the 1000 post limit [source](https://pushshift.io/).
2. Collect additional text information like comments from the individual reddit post
3. Explore more alternative social media platforms like twitter or facebook
